// src/components/Form/interface.ts
export type RuleType = "string" | "number" | "email" | "url" | "object";
// ...
export type ValidateMessage = string | (() => string);

type RangeMessages = {
  min?: ValidateMessage;
  max?: ValidateMessage;
  range?: ValidateMessage;
  len?: ValidateMessage;
};

export interface ValidateMessages {
  default?: ValidateMessage;
  required?: ValidateMessage;
  whitespace?: ValidateMessage;
  types?: Partial<Record<RuleType, ValidateMessage>>;
  string?: RangeMessages;
  number?: RangeMessages;
  array?: RangeMessages;
  pattern?: { mismatch?: ValidateMessage };
  [key: string]: any;
}